Move UuidCreateGenerator
